Add basic support for player position

+++ quartz/src/entities/mod.rs
@@ -4,17 +4,17 @@ pub mod player;
 
 
 pub struct Position {
-    pub x: f32,
-    pub y: f32,
-    pub z: f32,
+    pub x: f64,
+    pub y: f64,
+    pub z: f64,
 }
 
 impl From<BlockPosition> for Position {
     fn from(coord: BlockPosition) -> Self {
         Self {
-            x: coord.x as f32,
-            y: coord.y as f32,
-            z: coord.z as f32,
+            x: coord.x as f64,
+            y: coord.y as f64,
+            z: coord.z as f64,
         }
     }
 }
